chore: Move files to expected locations
