+++ src/Anichron.API/Security/JwtFactory.cs
@@ -1,4 +1,3 @@
-using Anichron.API.Infrastructure;
 using Anichron.API.Settings;
 using Anichron.Core.Domain;
 using Microsoft.Extensions.Options;
