+++ src/Anichron.API/Services/AuthService.cs
@@ -13,27 +13,6 @@ namespace Anichron.API.Services;
 public sealed record AuthTokens(string AccessToken, string RefreshToken);
 public sealed record AdminCreatedUser(Guid Id, string Username, string Email, string TemporaryPassword);
 
-public enum AuthError
-{
-    None = 0,
-    UsernameTaken = 1,
-    EmailTaken = 2,
-    InvalidCredentials = 3,
-    TokenInvalid = 4,
-    InvalidUsername = 5,
-    InvalidEmail = 6,
-    PasswordTooShort = 7,
-    PasswordTooLong = 8,
-    PasswordPwned = 9,
-    AccountDisabled = 10,
-    AccountTemporarilyLocked = 11,
-    InviteTokenInvalid = 12,
-    CannotModifySelf = 13,
-    UserNotFound = 14,
-    PathAlreadyAssigned = 15,
-    StorageConfigNotFound = 16,
-}
-
 public sealed record AuthResult<T>
 {
     public T? Value { get; init; }
